show flight days in inventory table
