+++ src/app/campaign/inventory/goal-form.component.scss
