+++ src/app/campaign/inventory/inventory-table.component.scss
@@ -5,17 +5,6 @@ $row-divider-height: 1px;
 $row-divider-border: $row-divider-height solid prx-theme-foreground(divider);
 $row-height: $row-line-height + $row-divider-height;
 
-:host {
-  margin: 20px 10px 20px 20px;
-}
-section {
-  margin: 20px 0;
-}
-.title {
-  color: prx-theme-foreground(text);
-  font-size: 1rem;
-  margin-bottom: 1rem;
-}
 .row {
   transition: background-color 0.6s cubic-bezier(0.25, 0.8, 0.25, 1);
   display: grid;
@@ -131,21 +120,8 @@ section {
   }
 }
 
-.row-highlight,
-.errors {
+.row-highlight {
   color: $warn;
-}
-
-.errors {
-  display: grid;
-  grid-row-gap: 0.5rem;
-  margin: 0 2rem 1rem 0;
-  font-style: italic;
-}
-
-.error {
-  display: flex;
-  align-items: center;
 }
 
 .preview {
